+++ e2e/main-nav.spec.ts
@@ -1,0 +1,13 @@
+import { test } from '@playwright/test';
+
+import { navLinks } from 'components/header';
+test('test main navigation', async ({ page }) => {
+  await page.goto('/');
+
+  const navigationBar = page.getByTestId('main-navigation');
+  for (const link of navLinks) {
+    await navigationBar.getByRole('link', { name: link.name }).click();
+    await page.waitForTimeout(2 * 1000);
+    await page.goto(link.href);
+  }
+});